REF: implement tests/indexes/objects/ (#31597)
